Create Collect_Local_Admins.ps1
Tanium Package to enumerate any users, local or domain, that is in the Local Administrators Group. It will write the results to "<Tanium_Install_Directory\AdQuery\LocalAdmins.dat", which can then be queried with a Sensor using the "Display_Local_Admins.ps1" script. It is recommended to spread this package deployment over time due to a small LDAP Query that is made each time the package is ran. You would then schedule the package to be reoccurring to keep the data up to date.