Create Display_Local_Admins.ps1
Tanium Sensor to display the Local Admin Data that was created from running the Package using the "Collect_Local_Admins.ps1" script. If the data file ("<Tanium_Install_Directory\AdQuery\LocalAdmins.dat") is present, it will display the contents. Otherwise, it will display not result. The output is to be configured to use the comma separated option for two columns. The first column should be titled "Computer Name" and the second "Account Name", both with the Result Type set to "Text".