Create Cylance_Token_Exsists.ps1
Tanium Sensor to enumerate if the Cylance Client has a Token or not. It will return a True or False accordingly. If a Token is not found, the system can't connect to a Cloud Console. It is also in an unprotected state once it has lost the Token and had a reboot. You can use the "Push_Cylance_Token.ps1" Package to push your company's Token to the Client.