Create Push_Cylance_Token.ps1
This Package can be used to push the Cylance InstallToken to the registry for any 32-bit or 64-bit system. For the Cylance Client to be able to ingest the "InstallToken" Key, the CylanceSvc Service must be running. If the Service is running, it will ingest the "InstallToken" Key and you will then see it disappear and the "LastStateRestorePoint" Key will then be present. If that doesn't happen after a short period of time, you may have a firewall issue where the Client can't communicate to the Cloud Console or a corrupt install and it needs to be reinstalled. You can use the "Cylance_Token_Exsists.ps1" Sensor to discover Systems that have lost or never recieved their InstallToken.